Reorganize to call *orceMesa only once.
The old logic was to force mesa, and then if an X server started up, we
would `unforce' mesa.  This did not seem to work reliably.  Now, we
figure out which we want, and then force it; this prevents setting and
unsetting in the SW case.  HW rendering seems to work more reliably
this way.


git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@5108 18c085ea-50e0-402c-830e-de6fd14e8384