Utilize new display abstractions.
This removes most of the code in the engine which was responsible for
creating and stopping X servers.  It is still the engine's job to
figure out which kind of display to utilize, but the initialization of
those displays is delegated to VisItDisplay.

This also fixes a small bug and a previously-missing comment in
XDisplay.C.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@5120 18c085ea-50e0-402c-830e-de6fd14e8384