Update Tuvok from it's repo, roughly version `0.6'.
This updates Tuvok to the bleeding edge version, which should
become 1.0 with few changes other than the DX renderer.

Updating the list of source files seemed painful.  This also
changes the makefile to just include another file with the list of
Tuvok sources.  The external file can then easily generated via a
find | awk or similar.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@6587 18c085ea-50e0-402c-830e-de6fd14e8384