Set an empty `CoreVolume' to the dataset.
Tuvok assumes it will always have a volume to work with, because
dataset instantiation implies reading a dataset from disk.  For
apps that won't use its LOD features, like VisIt, we need to
construct a fake volume so that Tuvok has something available;
we'll change/fill in the data for this volume later, at render
time.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@6606 18c085ea-50e0-402c-830e-de6fd14e8384