Allow a HW context even in serial mode.
This is a bit strange, as the only time the case comes up is when
the user is running a serial engine with -nowin.  We'll then
immediately pop up a window so we can render into it.

When scripting, this is basically equivalent to not using -nowin;
the difference is that we do the rendering on the engine in SR
mode.

Untested, but it probably avoids the annoying screen size issues
(viewer controls take up part of the window size specification)
because the window has no decorations, so it might be useful when
(e.g.) generating movies or some such.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@6609 18c085ea-50e0-402c-830e-de6fd14e8384