Update the version of GLEW (1.5.1-tjf).
This updates the version of GLEW to be my repository head, which
is essentially 1.5.1, a couple patches for issues I've brought up,
and then my patch series for allowing hard specification of a GL
library.

Expect a minimum of one more update.  Some work still needs to be
done for the mangled Mesa case, and then a GLEW release needs to
happen.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@6610 18c085ea-50e0-402c-830e-de6fd14e8384