Add a method to help with parallel debugging.
This adds a method, `PAR_WaitForDebugger' which spins until a
developer attaches a debugger to the process with MPI rank 0, and
sets the variable `i' to 0.  Code which calls it should never make
it into a release (or even a commit), but it's useful when you're
trying to debug a parallel problem.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@7174 18c085ea-50e0-402c-830e-de6fd14e8384