Added check to make sure we don't try to broadcast an empty vector.
I was hitting this with a test that used the Silo DB.  There's a
variety of submesh information that gets broadcast in
BroadcastGlobalInfo, and it seems like in some cases there's not
any submeshes so we don't want to broadcast any information at
all.  Unfortunately the check needs to be done in avtParallel,
because we don't know that *all* processors have no submeshes
until we do some global comm.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@7399 18c085ea-50e0-402c-830e-de6fd14e8384