Add a (very) simple program to rebaseline tests quickly.
This is a quick interface to run over our test images and rebase
them as quickly as possible.  The images are displayed one after
another, and a single hotkey gives a thumbs up or down and moves
on to the next image.

The images come from the previous night's davinci run.  Therefore,
this is not useful to rebase tests from local images.

There's still some (okay, a *lot* ;) rough edges in this version.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@8120 18c085ea-50e0-402c-830e-de6fd14e8384