Make sure HAVE_XLIB propagates to code that needs it.
We AC_SUBSTituted it, but never had a variable for that target in
make-variables or anywhere else.  Plotter/vtk is the only code
that seems to use HAVE_XLIB, so the makefile for it now gets
X_CPPFLAGS.

Further, in one of our mappers which relies on the define being
set on Linux, add an #else clause with an error, so we'll detect
that nothing is happening instead silently doing the wrong thing.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@8903 18c085ea-50e0-402c-830e-de6fd14e8384