Merge 2.0RC to trunk...
Fixed crashes in test suite for sessionview.py and simplekeyframe.py
where NULL drawable in avtActor object throws exceptions that get 
caught on engine way, way up in Executors.h, by passing a lot of 
code necessary to keep the engine in a useable state.

I replaced all instances of exceptions in avtActor with tests for
non-NULL and continuing execution as well as possible. 

avtPlot's ReleaseData method was ultimately the culprit setting
an avtActor's drawable to NULL. However, the avtActor object did
not appear to be designed to support a NULL drawable. So, the
two -- avtPlot setting avtActor's drawable to NULL and avtActor
throwing exceptions when its drawable is NULL -- seemed to be
inconsistent. I felt removing exceptions from avtActor was
more 'correct' removing the calls to set the drawable to NULL
in avtPlot's ReleaseData method. However, the latter approach
would have the intended effect as well.

I also replaced the one and only exception in VisWinAxesParallel.C
for similar reasons.


git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@10833 18c085ea-50e0-402c-830e-de6fd14e8384