Merge VisIt threading branch - svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/branches/camp/hp/src to truck.
This check in adds the changes needed to thread the filters of VisIt.
I still need to port the VTK changes to the new version of VTK 6 to have everything work.
I will send out a message when I have these changes.

By default VisIt is build in serial mode. You need to build VisIt with the threading option
to have a threaded version. This can be done with the --thread option to build visit script.
Or there is a CMake flag VISIT_THREAD.



git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@21206 18c085ea-50e0-402c-830e-de6fd14e8384