experimental ctest support
this patch is an initial pass at adding ctest support
to VisIt's regression suite. xml tags that describe test
results in the format that the cdash server expects have
been added to output log. The existing set of tests is
mapped on to a set of ctest test commands. Configure
VisIt with BUILD_TESTING ON to make use of it.



git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk@22772 18c085ea-50e0-402c-830e-de6fd14e8384