I added the -ppn (process per node) option to the JobSubmitter_mpiexec and JobSumbitter_qsub class. Now both the mpiexec and mpirun have the -ppn option.
I had a problem with the -osmesa flag not working on RedHat. The engine would still load the system 
libGL.so.1 library. This was happening because the -osmesa flag added the osmesa path to the end
of the LD_LIBRARY_PATH. I moved the osmesa path to the front of the list and everything works great.
I don't see any problem here. Only the libraries in that directory are our links to the osmesa library
and you are adding the -osmesa flag to make sure that they are loaded.



git-svn-id: svn+ssh://portal-auth.nersc.gov/project/projectdirs/visit/svn/visit/trunk@22982 18c085ea-50e0-402c-830e-de6fd14e8384