ctest support : threshold difference
added ability to use thresholded image difference test.
the test fails if absolute value of the difference at any
pixel is greater than the threshold. Added overides(threshold
needs to be greater than the default of 10) for 16 tests
failing on Apple, most of which failed because of differences
in line rendering between OSMesa and other GL.



git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23372 18c085ea-50e0-402c-830e-de6fd14e8384